Delete table, rows, or cols with backspace
Word table editing parity
