+++ blocks/edit/prose/index.js
@@ -1,25 +1,17 @@
 /* eslint-disable max-classes-per-file */
 import {
-  DOMParser,
   EditorState,
   EditorView,
-  TextSelection,
   history,
   buildKeymap,
   keymap,
   baseKeymap,
   tableEditing,
   columnResizing,
-  goToNextCell,
-  selectedRect,
-  isInTable,
-  addRowAfter,
   fixTables,
   liftListItem,
   sinkListItem,
   gapCursor,
-  InputRule,
-  inputRules,
   Y,
   WebsocketProvider,
   ySyncPlugin,
@@ -39,6 +31,7 @@ import base64Uploader from './plugins/base64uploader.js';
 import { COLLAB_ORIGIN, getDaAdmin } from '../../shared/constants.js';
 import { getLocClass } from './loc-utils.js';
 import { getSchema } from './schema.js';
+import { handleTableBackspace, handleTableTab, getEnterInputRulesPlugin } from './plugins/keyHandlers.js';
 
 const DA_ORIGIN = getDaAdmin();
 
@@ -149,36 +142,6 @@ function generateColor(name, hRange = [0, 360], sRange = [60, 80], lRange = [40,
   return `#${f(0)}${f(8)}${f(4)}`;
 }
 
-export function getDashesInputRule() {
-  return new InputRule(
-    /^---[\n]$/,
-    (state, match, start, end) => {
-      const div = document.createElement('div');
-      div.append(document.createElement('hr'));
-      const newNodes = DOMParser.fromSchema(state.schema).parse(div);
-
-      const selection = TextSelection.create(state.doc, start, end);
-      dispatchTransaction(state.tr.setSelection(selection).replaceSelectionWith(newNodes));
-    },
-  );
-}
-
-// This function returns a modified inputrule plugin that triggers when the regex in the
-// rule matches and the Enter key is pressed
-export function getEnterInputRulesPlugin() {
-  const irsplugin = inputRules({ rules: [getDashesInputRule()] });
-
-  const hkd = (view, event) => {
-    if (event.key !== 'Enter') return false;
-    const { $cursor } = view.state.selection;
-    if ($cursor) return irsplugin.props.handleTextInput(view, $cursor.pos, $cursor.pos, '\n');
-    return false;
-  };
-  irsplugin.props.handleKeyDown = hkd; // Add the handleKeyDown function
-
-  return irsplugin;
-}
-
 export default function initProse({ editor, path }) {
   // Destroy ProseMirror if it already exists - GH-212
   if (window.view) delete window.view;
@@ -226,25 +189,6 @@ export default function initProse({ editor, path }) {
     );
   }
 
-  const handleTableTab = (direction) => {
-    const isCursorInLastTableCell = (rect) => {
-      const { left, bottom } = rect;
-      const { height, width } = rect.map;
-      return left + 1 === width && bottom === height;
-    };
-
-    const gtnc = goToNextCell(direction);
-    return (state, dispatch) => {
-      if (!isInTable(state)) return false;
-      const rect = selectedRect(state);
-      if (isCursorInLastTableCell(rect)) {
-        addRowAfter(state, dispatch);
-        return gtnc(window.view.state, dispatch);
-      }
-      return gtnc(state, dispatch);
-    };
-  };
-
   let state = EditorState.create({
     schema,
     plugins: [
@@ -261,9 +205,9 @@ export default function initProse({ editor, path }) {
       sectionPasteHandler(schema),
       base64Uploader(schema),
       columnResizing(),
-      tableEditing(),
-      getEnterInputRulesPlugin(),
+      getEnterInputRulesPlugin(dispatchTransaction),
       keymap(buildKeymap(schema)),
+      keymap({ Backspace: handleTableBackspace }),
       keymap(baseKeymap),
       keymap({
         'Mod-z': yUndo,
@@ -279,6 +223,7 @@ export default function initProse({ editor, path }) {
         'Shift-Tab': liftListItem(schema.nodes.list_item),
       }),
       gapCursor(),
+      tableEditing(),
       history(),
     ],
   });
